third_party/WebKit/LayoutTests/fast/dom/HTMLScriptElement/append-child-adopt-node-error-crash-expected.txt

CONSOLE MESSAGE: line 9: PASS unless crash.
 
